+++ server/src/main/java/org/opensearch/client/Requests.java
@@ -49,6 +49,7 @@ import org.opensearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest
 import org.opensearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
 import org.opensearch.action.admin.cluster.shards.routing.weighted.get.ClusterGetWeightedRoutingRequest;
 import org.opensearch.action.admin.cluster.shards.routing.weighted.put.ClusterPutWeightedRoutingRequest;
+import org.opensearch.action.admin.cluster.shards.routing.weighted.delete.ClusterDeleteWeightedRoutingRequest;
 import org.opensearch.action.admin.cluster.snapshots.create.CreateSnapshotRequest;
 import org.opensearch.action.admin.cluster.snapshots.delete.DeleteSnapshotRequest;
 import org.opensearch.action.admin.cluster.snapshots.get.GetSnapshotsRequest;
@@ -567,5 +568,14 @@ public class Requests {
      */
     public static ClusterGetWeightedRoutingRequest getWeightedRoutingRequest(String attributeName) {
         return new ClusterGetWeightedRoutingRequest(attributeName);
+    }
+
+    /**
+     * Deletes weights for weighted round-robin search routing policy
+     *
+     * @return delete weight request
+     */
+    public static ClusterDeleteWeightedRoutingRequest deleteWeightedRoutingRequest() {
+        return new ClusterDeleteWeightedRoutingRequest();
     }
 }